% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{BackwardLinear}
\alias{BackwardLinear}
\title{Autoregressive Backward Algorithm adapted from Stanculescu et al}
\usage{
BackwardLinear(
  algae_data,
  toxin_data,
  time,
  tran,
  mu_a,
  k_a,
  tox_em,
  states,
  toxin_states
)
}
\arguments{
\item{algae_data}{Algae count data}

\item{toxin_data}{Toxin data measured in micro grams / OA equivalent}

\item{time}{Day to calculate forward probability at}

\item{tran}{Transition Markov state probabilities}

\item{mu_a}{mean negative binomial parameter}

\item{k_a}{size negative binomial parameter}

\item{tox_em}{Classification matrix from ordinal regression parameters}
}
\value{
Return backward probabilities from day time to end of data
}
\description{
Calculates backward probabilities accounting for autoregressive dependence in the toxin measurements
}
