% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{BackwardLinearHelper}
\alias{BackwardLinearHelper}
\title{Part of backward algorithm}
\usage{
BackwardLinearHelper(
  new_beta,
  alg_data,
  tox_data,
  obs_tox,
  tran,
  mu_a,
  k_a,
  tox_em
)
}
\description{
Part of backward algorithm
}
