% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{BackwardLinearHelperMissing}
\alias{BackwardLinearHelperMissing}
\title{Part of backward algorithm}
\usage{
BackwardLinearHelperMissing(
  new_beta,
  alg_data,
  obs_tox,
  tran,
  mu_a,
  k_a,
  tox_em
)
}
\description{
Part of backward algorithm
}
