% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{CalcBetas}
\alias{CalcBetas}
\title{Calculates ordinal logistic regression parameters}
\usage{
CalcBetas(algae_data, toxin_data, forw, backw, tran, mu_a, k, tox_em, denom)
}
\arguments{
\item{algae_data}{Algae count data}

\item{toxin_data}{Toxin data measured in micro grams / OA equivalent}

\item{forw}{Forward probabilities}

\item{backw}{Backward probabilities}

\item{tran}{Old Markov state transition probabilities}

\item{mu_a}{mean negative binomial parameter}

\item{k}{size negative binomial parameter}

\item{tox_em}{Classification matrix from ordinal regression parameters}

\item{denom}{Denominator used in expected value calculations, total likelihood of all data}
}
\value{
Returns betas for ordinal logistic regression
}
\description{
Calculates ordinal logistic regression parameters
}
