% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{CalcInit}
\alias{CalcInit}
\title{Calculates latent state initial probabilities}
\usage{
CalcInit(forw, backw, states, toxin_states)
}
\arguments{
\item{forw}{Forward probabilities}

\item{backw}{Backward probabilities}
}
\value{
Returns initial probabilities
}
\description{
Calculates latent state initial probabilities
}
