% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ClassificationAlgae}
\alias{ClassificationAlgae}
\title{Calculates probability of observed algae}
\usage{
ClassificationAlgae(observed_val, latent_val, mu_a, k_a)
}
\arguments{
\item{observed_val}{Observed algae measurement}

\item{latent_val}{Latent Markov state}

\item{mu_a}{mean negative binomial parameter}

\item{k_a}{size negative binomial parameter}
}
\value{
Probability of observed algae
}
\description{
Calculates probability of observed algae based on latent state and negative binomial parameters
}
