% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ClassificationToxin}
\alias{ClassificationToxin}
\title{Calculates probability of observed toxin}
\usage{
ClassificationToxin(observed_val, last_observed_val, latent_val, tox_em)
}
\arguments{
\item{observed_val}{Observed toxin measurement}

\item{latent_val}{Latent Markov state}

\item{tox_em}{Classification matrix from ordinal regression parameters}
}
\value{
Probability of observed toxin
}
\description{
Calculates probability of observed toxin based on latent state and ordinal regression parameters
}
