% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{CollapseBackw}
\alias{CollapseBackw}
\title{Collapses backward probabilities}
\usage{
CollapseBackw(backw, time)
}
\arguments{
\item{backw}{Backward probabilities to collapse}

\item{time}{Time to collapse at}
}
\value{
Returns collapsed backward quantity
}
\description{
Collapses backward probabilities over missing data
}
