% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{CollapseForw}
\alias{CollapseForw}
\title{Collapses forward probabilities}
\usage{
CollapseForw(forw, time)
}
\arguments{
\item{forw}{Forward probabilities to collapse}

\item{time}{Time to collapse at}
}
\value{
Returns collapsed forward quantity
}
\description{
Collapses forward probabilities over missing data
}
