% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{Cont2Ord}
\alias{Cont2Ord}
\title{Continuous to discrete for toxin data}
\usage{
Cont2Ord(toxin_data, co1, co2, co3)
}
\arguments{
\item{toxin_data}{Toxin data measured in micro grams / OA equivalent}

\item{co1}{Cutoff 1 for toxin data}

\item{co2}{Cutoff 1 for toxin data}

\item{co3}{Cutoff 1 for toxin data}
}
\value{
Returns discretized toxin data
}
\description{
Continuous to discrete for toxin data
}
