% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ForwardLinear}
\alias{ForwardLinear}
\title{Autoregressive Forward Algorithm adapted from Stanculescu et al}
\usage{
ForwardLinear(algae_data, toxin_data, time, init, tran, mu_a, k_a, tox_em)
}
\arguments{
\item{algae_data}{Algae count data}

\item{toxin_data}{Toxin data measured in micro grams / OA equivalent}

\item{time}{Day to calculate forward probability at}

\item{init}{Initial Markov state probabilities}

\item{tran}{Transition Markov state probabilities}

\item{mu_a}{mean negative binomial parameter}

\item{k_a}{size negative binomial parameter}

\item{tox_em}{Classification matrix from ordinal regression parameters}
}
\value{
Return forward probabilities from day 1 to day time(variable name)
}
\description{
Calculates forward probabilities accounting for autoregressive dependence in the toxin measurements
}
