% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ForwardLinearHelper}
\alias{ForwardLinearHelper}
\title{Part of forward algorithm}
\usage{
ForwardLinearHelper(
  old_alpha,
  alg_data,
  tox_data,
  last_tox,
  tran,
  mu_a,
  k_a,
  tox_em
)
}
\description{
Part of forward algorithm
}
