% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ForwardLinearHelperMissing}
\alias{ForwardLinearHelperMissing}
\title{Part of forward algorithm}
\usage{
ForwardLinearHelperMissing(
  old_alpha_mat,
  alg_data,
  tox_data,
  tran,
  mu_a,
  k_a,
  tox_em
)
}
\description{
Part of forward algorithm
}
