% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{GenerateSimulatedMC}
\alias{GenerateSimulatedMC}
\title{Generates simulated Markov chain data}
\usage{
GenerateSimulatedMC(init, tran, toxin_data_len, mu_a, k, tox_em, missing_perc)
}
\arguments{
\item{init}{Initial Markov state probabilities}

\item{tran}{Transition Markov state probabilities}

\item{toxin_data_len}{Length of days to simulate}

\item{mu_a}{mean negative binomial parameter}

\item{k}{size negative binomial parameter}

\item{tox_em}{Classification matrix from ordinal regression parameters}

\item{missing_perc}{Percent of data to remove from simulated data}
}
\value{
Returns simulated data
}
\description{
Generates simulated Markov chain data
}
