% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{K2Phi}
\alias{K2Phi}
\title{Transforms negative binomial size parameter to phi}
\usage{
K2Phi(mu, k_a)
}
\arguments{
\item{mu}{mean paramater}

\item{k_a}{size parameter}
}
\value{
Returns phi paramaeter for negative binomial
}
\description{
Transforms negative binomial size parameter to phi parameter. Opposite of Phi2K
}
