% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{LogLikenbinom}
\alias{LogLikenbinom}
\title{Calculates log likelihood of negative binomial for algae data}
\usage{
LogLikenbinom(x, data, weights)
}
\arguments{
\item{x}{vector of mean and size parameter}

\item{data}{Algae data to calculate likelihood of}

\item{weights}{Weights from previous function ProbWeights}
}
\value{
Returns log likelihood for negative binomial of algae data
}
\description{
Calculates log likelihood of negative binomial for algae data
}
