% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{Ord2Mat}
\alias{Ord2Mat}
\title{Transforms ordinal regression parameters to classification matrix}
\usage{
Ord2Mat(coeffs, zetas, states, toxin_states)
}
\arguments{
\item{coeffs}{Regression coefficents}

\item{zetas}{Regression intercepts}
}
\value{
Returns classification matrix
}
\description{
Transforms ordinal regression parameters to classification matrix
}
