% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{Phi2K}
\alias{Phi2K}
\title{Transforms negative binomial phi parameter to size}
\usage{
Phi2K(mu, phi)
}
\arguments{
\item{mu}{mean paramater}

\item{phi}{other parameterization parameter}
}
\value{
Returns size paramaeter for negative binomial
}
\description{
Transforms negative binomial phi parameter to size parameter. There are two different possible parameterizations of negative binomial
Size (k) is easier to interpret but R uses phi
}
