% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ProbWeights}
\alias{ProbWeights}
\title{Calculates relative weights of the binary Markov states}
\usage{
ProbWeights(data, forw, backw)
}
\arguments{
\item{data}{Either algae or toxin data to get the length of time in study}

\item{forw}{Forward probabilities}

\item{backw}{Backward probabilities}
}
\value{
Returns relative weights of Markov latent states
}
\description{
Calculates relative weights of the binary Markov states
}
