% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{RunEM}
\alias{RunEM}
\title{Runs EM}
\usage{
RunEM(algae_data, toxin_data, init, tran, mu_a, k_a, betas, threshold)
}
\arguments{
\item{algae_data}{Algae count data}

\item{toxin_data}{Toxin data measured in micro grams / OA equivalent}

\item{init}{Initial Markov state probabilities}

\item{tran}{Old Markov state transition probabilities}

\item{mu_a}{mean negative binomial parameter}

\item{k_a}{size negative binomial parameter}

\item{betas}{Ordinal logistic regression coefficients}

\item{threshold}{Threshold for algae data. Values lower than threshold are set to 0}
}
\value{
Returns initial probabilities
}
\description{
Runs EM algorithm and outputs estimated parameters
}
