% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{SimData}
\alias{SimData}
\title{Generates simulated data}
\usage{
SimData(
  init_true,
  tran_true,
  mu_a_true,
  k_true,
  betas_true,
  missing_perc,
  data_len,
  states,
  toxin_states
)
}
\arguments{
\item{init_true}{Initial Markov state probabilities for simulated data}

\item{tran_true}{Transition Markov state probabilities for simulated data}

\item{mu_a_true}{mean negative binomial parameter for simulated data}

\item{k_true}{size negative binomial parameter}

\item{betas_true}{Betas for simulated data}

\item{missing_perc}{Percent of data to remove from simulated data}

\item{data_len}{Number of days to simulate}
}
\value{
Returns simulated data
}
\description{
Generates simulated data
}
