% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BivariateAutoregressiveHMM.R
\name{ProbSt}
\alias{ProbSt}
\title{Calculates latent state probability at specific time}
\usage{
ProbSt(forw, backw, time, states, toxin_states)
}
\arguments{
\item{forw}{Forward probabilities}

\item{backw}{Backward probabilities}

\item{time}{Time to calculate Markov at}
}
\value{
Returns latent state probability at specific time
}
\description{
Calculates latent state probability at specific time
}
